+++ examples/7_resolve_recursively.py
@@ -1,10 +1,9 @@
 from __future__ import annotations
 import asyncio
-from random import random
 from typing import Optional
 from time import time
 from pydantic import BaseModel
-from pydantic_resolve import resolve
+from pydantic_resolve import Resolver
 
 t = time()
 
@@ -24,7 +23,7 @@ class A(BaseModel):
 
 async def main():
     a = A()
-    result = await resolve(a)
+    result = await Resolver().resolve(a)
     print(result.json())
     print(f'total {time() - t}')